fix: #198 게시글 수정 페이지 반응형 작업 및 textarea->input 교체
